requirements.txt
langchain==0.3.12
langchain-core==0.3.45
langchain-community==0.3.12
langchain-openai==0.3.8

langgraph==0.2.70
langgraph-checkpoint==2.1.2
langgraph-checkpoint-sqlite==2.0.11

langsmith==0.1.147

openai==1.60.0
tiktoken==0.7.0

httpx==0.27.2
aiohttp==3.9.5

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2

transformers==4.44.2
sentence-transformers==3.0.1
huggingface-hub==0.24.6

sqlite-vec==0.1.9
streamlit==1.38.0

pydantic==2.9.2
python-dotenv==1.0.1
tenacity==8.3.0